+++ CHANGELOG.md
@@ -6,6 +6,19 @@ and this module adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [5.1.1] - 2025-07-15
+
+### Added
+
+- Support for PremiumV2_LRS & UltraSSD_LRS storage account type (data disk).
+- Validations for the PremiumV2_LRS & UltraSSD_LRS data disks, Caching, On Demand Bursting Enabled, Write Acceleration, Disks Iops & Disks Mbps.
+- Variable "zones" & "sku" to the public_ip_config block with its validations.
+- "os_publisher", "os_offer" & "additional_capabilities" to the virtual_machine_config" block.
+
+### Changed
+
+- In newer Azure provider versions, sku in public_ip_config defaults to "Standard" instead of "Basic". As a result, this optional variable was explicitly added with the default value set to "Standard"
+
 ## [5.1.0] - 2025-02-27
 
 ### Added